Fix sizing of sponsor logos
Every sponsor has about the same size of their logo on screen now rather
than logos with certain dimensions being rendered bigger than others.

When logos included transparent spacing in the images I've made an
exception to give them more spacing so the actual visible part of the
logo is about the same size as the others.

Also tweaked the rest of the CSS to render better on devices with
smaller screens to have the same behavior.