Fix schedule alignment
This commit makes sure that all the times in the table are aligned under
one another in different rows. This makes the schedule more easy to
read, because the times aren't in a different position per row.

Also make sure every column doesn't use more space than it needs to and
has enough space on mobile.