Fix schedule table on mobile Safari
Fix the schedule tables on mobile Safari by setting the width property
to `auto` so it will not try to cram the `80%` width into the event
column.

On other browsers you have a bit more padding on the right side of the
table cell with the time now, but that is hardly noticed and may be
fixed later with the `box-sizing: border-box` property set on all
elements.