Remove horizontal scroll from website
A horizontal scroll was noticeable on smaller screens, leaving a white
gap on the right side of the website all the time if you scroll too far
right.

This change fixes that by changing the `box-sizing` property to
`border-box`. This makes the sizing properties behave more like expected
making padding not account for the width of an element.

Downside is that all rules were previously setup with with the other
`box-sizing` rules in mind and thought they had more space (the white
gap on the right side), so everything is a bit smaller now that relied
on auto sizing based on page width.

Also the sponsor page has this huge table with all the packages which
cannot be rendered within the current layout on mobile. We may need to
move that to a separate page for mobile, put the table in a partial and
render it normally on desktop.