Update logos and spacing in sponsor section
Make sure logos are properly aligned and have enough space around them
to stand out. Adjust some new logos sizes to make sure they are about as
large as the rest of the logos.

Logos no longer only use the space their images are, but always have 50%
of the row on larger screens making sure we have a two column layout.
This makes the list a bit more predictable and easier on the eye as some
rows don't have 3 or 4 items while others have two.