Improve sponsor logo rendering
With the image tag and all the logos in different sizes it's hard to
come up with good rules that would apply to all logos.

Instead remove the image tag and apply the logos as background images.
This makes it easier to fit the logos in the allocated space while
retaining their aspect ratios without custom rules for every logo.

Still added some custom rules for certain sized logos to avoid them
using too much empty space for the `a`-tag.