Only load social icons once
The GitHub and Twitter SVG icons were included for every time they were
used. This change uses references to point to one SVG icon for GitHub
and Twitter. This clean up the view considerably and avoid mistakes with
copy-pasting icons.

Also replaced the Twitter icon in the footer with the reference.

Use `tw-icon` and `gh-icon` as IDs so that they don't get blocked by
adblockers. It's just an icon. Also fixes the issue that the Twitter
icon in the footer doesn't currently load on mobile Safari with Purify.

Use both the `href` and `xlink:href` attributes to load the icons so we
still support some older browsers.